add MesosSchedulerBackendUtilSuite.scala
